+++ src/methods/changeContext.js
@@ -3,12 +3,21 @@
 function changeContext ( dependencies, state ) {
 const 
           { 
-               shortcuts
-             , listenOptions 
-             , currentContext
+                shortcuts
+              , currentContext
           } = state
         , { ev } = dependencies
         ;
+
+
+
+function expose () {
+                ev.on ( '*', (...args) => {
+                                if ( state.exposeShortcut )   state.exposeShortcut ( ...args )
+                        })
+        } // expose func.
+
+
 
 /**
  * @function changeContext
@@ -19,14 +28,6 @@ const
 return function changeContext ( contextName = false ) {
         const current = currentContext.name;
         
-        listenOptions.maxSequence = 1
-        listenOptions.maxClicks = 1
-        
-        if ( listenOptions.keyIgnore >= 0 ) {   
-                        clearTimeout ( listenOptions.keyIgnore )
-                        listenOptions.keyIgnore = null
-                }
-
         if ( !contextName ) {   // Switch off all shortcuts if contextName is not defined
                         ev.reset ()
                         currentContext.name = null
@@ -34,26 +35,19 @@ return function changeContext ( contextName = false ) {
                 }
         if ( current === contextName ) return   // Do nothing if contextName is the same as current
         if ( !shortcuts [ contextName ] ) {   // If contextName is not defined
-                        ev.emit ( 'shortcuts-error', `Context '${ contextName }' does not exist` )
+                        ev.emit ( '@shortcuts-error', `Context '${ contextName }' does not exist` )
                         return
                 }
         if ( shortcuts[current] ) {
                         ev.reset ()   // Disable all shortcuts from current context
                 }
-        Object.entries ( shortcuts [ contextName ]).forEach ( ([shortcutName, list ]) => {   // Enable new context shortcuts and set a listenOptions 'maxSequence' and 'maxClicks'       
-                        let isMouseEv = shortcutName.includes ( 'MOUSE-CLICK-' );
-                        if ( isMouseEv ) {   // Set mouse max clicks
-                                        let [ , , , count ] = shortcutName.split('-')
-                                        let c = parseInt ( count );   // Number of clicks
-                                        if ( listenOptions.maxClicks < c )   listenOptions.maxClicks = c
-                                }
-                        else {   // Set key max sequence
-                                        let sequenceArraySize = shortcutName.split(',').length;
-                                        if ( listenOptions.maxSequence < sequenceArraySize )   listenOptions.maxSequence = sequenceArraySize
-                                }
-                        list.forEach ( fn => ev.on ( shortcutName, fn )   )    // Enable new context shortcuts
-                })   
-        currentContext.name = contextName        
+
+        currentContext.name = contextName 
+        state.plugins.forEach ( plugin => plugin.contextChange ( contextName )    )        // Inform plugins for context change
+        Object.entries ( shortcuts[contextName] ).forEach ( ([shortcutName, list ]) => {   // Enable new context shortcuts
+                        list.forEach ( fn => ev.on ( shortcutName, fn )    )
+                })
+        expose ()
 }} // changeContext func.