+++ src/methods/index.js
@@ -1,9 +1,6 @@
-import _findTarget      from './_findTarget.js'
-import _listen         from './_listen.js'
-import _readKeyEvent   from './_readKeyEvent.js'
-import _readMouseEvent from './_readMouseEvent.js'
-import _readShortcut   from './_readShortcut.js'
-import _specialChars   from './_specialChars.js'
+import _normalizeWithPlugins    from './_normalizeWithPlugins.js'
+import _readShortcutWithPlugins from './_readShortcutWithPlugins.js'
+import _systemAction            from './_systemAction.js'
 
 import load           from './load.js'
 import unload         from './unload.js'
@@ -14,13 +11,10 @@ import listShortcuts  from './listShortcuts.js'
 
 export default {
 // Internal methods
-    _findTarget
-  , _listen
-  , _readKeyEvent
-  , _readMouseEvent
-  , _readShortcut
-  , _specialChars
-
+   _normalizeWithPlugins
+ , _readShortcutWithPlugins
+ , _systemAction
+    
 // Public methods
   , changeContext
   , listShortcuts